Split scope tests
