+++ src/scope/rootScope.ts
@@ -4,7 +4,7 @@ import ScopeImpl from './scopeImpl';
 import ScopeIterator from './scopeIterator';
 
 export default class RootScope extends ScopeIterator {
-  *scopes(events: Generator<Event>): Generator<Scope> {
+  *scopes(events: IterableIterator<Event>): Generator<Scope> {
     for (const event of events) {
       if (event.isCall() && !event.parent) {
         yield new ScopeImpl(event);
